+++ internal/config/config_test.go
@@ -39,9 +39,9 @@ func TestDefaultConfig(t *testing.T) {
 		t.Error("expect no lastfm secret")
 	}
 
-	if len(config.Music.RadioStreams) == 0 {
-		t.Error("expect radio streams")
-	}
+	// if len(config.Music.RadioStreams) == 0 {
+	// 	t.Error("expect radio streams")
+	// }
 	if len(config.Podcast.Series) == 0 {
 		t.Error("expect podcast streams")
 	}