+++ takeout.go
@@ -19,6 +19,6 @@ package takeout
 
 const (
 	AppName = "Takeout"
-	Version = "0.13.2"
+	Version = "0.14.0"
 	Contact = "takeoutfm.com"
 )